% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.r
\name{calculate_actions_per_AU}
\alias{calculate_actions_per_AU}
\title{Calculate average actions per WAU for each user group, subaggregate, and aggregate,
and for each date range.}
\usage{
calculate_actions_per_AU(long_flash_report_3, long_flash_report_total_actions)
}
\arguments{
\item{long_flash_report_3}{The rbinded results of
summarise_by_subaggregate, summarise_in_aggregate, and curate_user_groups.}

\item{long_flash_report_total_actions}{The result of calling
calculate_total_actions.}
}
\value{
A data frame with the same column structure as the input, but with
additional rows corresponding to the actions_per_AU variable.
}
\description{
Calculate average actions per WAU for each user group, subaggregate, and aggregate,
and for each date range.
}

