% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.r
\name{calculate_NRR}
\alias{calculate_NRR}
\title{Calculate notifications_response_rate for each user group, subaggregate, and aggregate,
and for each date range.}
\usage{
calculate_NRR(long_flash_report_3)
}
\arguments{
\item{long_flash_report_3}{The rbinded results of
summarise_by_subaggregate, summarise_in_aggregate, and curate_user_groups.}
}
\value{
A data frame with the same column structure as the input, but with
additional rows corresponding to the notifications_response_rate variable.
}
\description{
Calculate notifications_response_rate for each user group, subaggregate, and aggregate,
and for each date range.
}

