% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.r
\name{calculate_total_actions}
\alias{calculate_total_actions}
\title{Calculate total actions for each user group, subaggregate, and aggregate,
and for each date range.}
\usage{
calculate_total_actions(long_flash_report_3)
}
\arguments{
\item{long_flash_report_3}{The rbinded results of
summarise_by_subaggregate, summarise_in_aggregate, and curate_user_groups.}
}
\value{
A data frame with the same column structure as the input, but with
additional rows corresponding to the total actions variable.
}
\description{
Calculate total actions for each user group, subaggregate, and aggregate,
and for each date range.
}

