% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.r
\name{calculate_WAU_percentage}
\alias{calculate_WAU_percentage}
\title{Calculate WAU percentage for each user group, subaggregate, and aggregate,
and for each date range.}
\usage{
calculate_WAU_percentage(long_flash_report_3)
}
\arguments{
\item{long_flash_report_3}{The rbinded results of
summarise_by_subaggregate, summarise_in_aggregate, and curate_user_groups.}
}
\value{
A data frame with the same column structure as the input, but with
additional rows corresponding to the WAU percent variable.
}
\description{
Calculate WAU percentage for each user group, subaggregate, and aggregate,
and for each date range.
}

