% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.r
\name{curate_user_groups}
\alias{curate_user_groups}
\title{Removes undesireable user groups from the final report.}
\usage{
curate_user_groups(long_flash_report_dates_formatted)
}
\arguments{
\item{long_flash_report_dates_formatted}{The result of calling
format_LFR_dates.}
}
\value{
A data frame equivalent to the input, except that certain
user_groups have been filtered out.
}
\description{
Removes undesireable user groups from the final report.
}

