% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripting_functions.r
\name{define_option_list}
\alias{define_option_list}
\title{Defines the option list for the master script to call.}
\usage{
define_option_list()
}
\value{
A list with entries host, port, user, pass, rundate,
minweek, maxweek, yearbeginning, outloc, outname
}
\description{
Defines the option list for the master script to call.
}

