% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.r
\docType{class}
\name{FlashReportQuery-class}
\alias{FlashReportQuery-class}
\title{An S4 class to represent a Flash Report query.}
\description{
An S4 class to represent a Flash Report query.
}
\section{Slots}{

\describe{
\item{\code{min_date}}{A Date object, representing the first day of the date ranges
for which results are collected.}

\item{\code{max_date}}{A Date object, representing the final day of the date ranges
for which results are collected.}

\item{\code{user_group}}{integer vector of user_ids to restrict results to. Cannot 
specify both user_group and user_group_query.}

\item{\code{user_group_query}}{Character containing a SQL query that returns a
table with a single column, user_id. Cannot specify both user_group and
user_group_query.}

\item{\code{user_group_name}}{Character, a name for the user group you're running 
the query on.}

\item{\code{query_prototype}}{A character. Represents the SQL query that will return
the desired results, but the date filters are general, and need to be subbed
in using the 'substitute_dates' method.}

\item{\code{query}}{A character. The actual query to run in Redshift.}

\item{\code{raw_results}}{A data frame. The results of running "query".}

\item{\code{final_results}}{A data frame. The formatted results that are ready to be
bound to all other query results.}
}}

