% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.r
\name{format_LFR_dates}
\alias{format_LFR_dates}
\title{Formats dates in the results of get_results to a specified standard.}
\usage{
format_LFR_dates(long_flash_report)
}
\arguments{
\item{long_flash_report}{The result of calling get_results.}
}
\value{
A data frame equivalent to the input, except that the dates are in a
different format.
}
\description{
Formats dates in the results of get_results to a specified standard.
}

