% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.r
\docType{methods}
\name{format_raw_results}
\alias{format_raw_results}
\alias{format_raw_results,auQuery-method}
\alias{format_raw_results,notificationsQuery-method}
\alias{format_raw_results,paQuery-method}
\title{A generic function to format the raw_results slot of a FlashReportQuery
object into a standardized format that can be rbinded with all other
results.}
\usage{
format_raw_results(frq)

\S4method{format_raw_results}{auQuery}(frq)

\S4method{format_raw_results}{paQuery}(frq)

\S4method{format_raw_results}{notificationsQuery}(frq)
}
\arguments{
\item{frq}{An object that inherits from FlashReportQuery.}
}
\value{
A FlashReportQuery object of the same subtype that was input. The
final_results slot contains a data.frame with columns (user_group,
date_range, variable, value).
}
\description{
A generic function to format the raw_results slot of a FlashReportQuery
object into a standardized format that can be rbinded with all other
results.
}
\section{Methods (by class)}{
\itemize{
\item \code{auQuery}: Format results of an Active Users query.

\item \code{paQuery}: Format results of a platform action query.

\item \code{notificationsQuery}: Format results of a notifications query.
}}

