% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.r
\docType{methods}
\name{get_prototype}
\alias{get_prototype}
\alias{get_prototype,auQuery-method}
\alias{get_prototype,notificationsQuery-method}
\alias{get_prototype,paQuery-method}
\title{A generic function to grab the correct query prototype and stick it into the
query_prototype slot of a FlashReport query.}
\usage{
get_prototype(frq)

\S4method{get_prototype}{auQuery}(frq)

\S4method{get_prototype}{paQuery}(frq)

\S4method{get_prototype}{notificationsQuery}(frq)
}
\arguments{
\item{frq}{An object that inherits from FlashReportQuery.}
}
\value{
A FlashReportQuery object of the same subtype that was input.
}
\description{
A generic function to grab the correct query prototype and stick it into the
query_prototype slot of a FlashReport query.
}
\section{Methods (by class)}{
\itemize{
\item \code{auQuery}: Get the query prototype for an active users query.

\item \code{paQuery}: Get the query prototype for a platform actions 
query.

\item \code{notificationsQuery}: Get the query prototype for a notifications query.
}}

