% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{gloo_internal_emails}
\alias{gloo_internal_emails}
\title{A string containing all internal emails for Gloo employees.}
\format{A length-one character vector.}
\usage{
gloo_internal_emails
}
\description{
A string containing all internal emails for Gloo employees.
}
\keyword{datasets}

