% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{isFL}
\alias{isFL}
\title{A data.frame containing all (user_group, isFL) pairs.}
\format{A length-one character vector.}
\usage{
isFL
}
\description{
A data.frame containing all (user_group, isFL) pairs.
}
\keyword{datasets}

