% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{pa_classes}
\alias{pa_classes}
\title{A string containing all (platform_action, flash_report_category) pairs.}
\format{A length-one character vector.}
\usage{
pa_classes
}
\description{
A string containing all (platform_action, flash_report_category) pairs.
}
\keyword{datasets}

