% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{query_pa_flash_cat}
\alias{query_pa_flash_cat}
\title{A generic query to create a temporary table in Redshift that matches
platform actions to their Flash Report Categories.}
\format{string}
\usage{
query_pa_flash_cat
}
\description{
A generic query to create a temporary table in Redshift that matches
platform actions to their Flash Report Categories.
}
\keyword{datasets}

