% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{query_pa_flash_cat_insert}
\alias{query_pa_flash_cat_insert}
\title{A helper query to insert the proper values into the pa_flash_cat temporary
table.}
\format{string}
\usage{
query_pa_flash_cat_insert
}
\description{
A helper query to insert the proper values into the pa_flash_cat temporary
table.
}
\keyword{datasets}

