% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{query_prototype_list}
\alias{query_prototype_list}
\title{A list of strings containing all query prototypes.}
\format{A list of strings}
\usage{
query_prototype_list
}
\description{
A list of strings containing all query prototypes.
}
\keyword{datasets}

