% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{query_user_flash_cat_sub}
\alias{query_user_flash_cat_sub}
\title{A primitive version fo query_user_flash_cat that contains a string
xyz_gloo_internal_email_xyz which can be subbed out for a given list of
email addresses.}
\format{string}
\usage{
query_user_flash_cat_sub
}
\description{
A primitive version fo query_user_flash_cat that contains a string
xyz_gloo_internal_email_xyz which can be subbed out for a given list of
email addresses.
}
\keyword{datasets}

