% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.r
\docType{methods}
\name{run_query}
\alias{run_query}
\alias{run_query,FlashReportQuery-method}
\title{A generic function to run the query in the query slot of a FlashReportQuery
object, and return the results in the raw_results slot.}
\usage{
run_query(frq, ...)

\S4method{run_query}{FlashReportQuery}(frq, con = redshift_connection$con)
}
\arguments{
\item{frq}{A FlashReportQuery object.}

\item{...}{Placeholder for database connection object.}

\item{con}{The database connection object to run the query through. Defaults
to insightsdev.}
}
\value{
A FlashReportQuery object.
}
\description{
A generic function to run the query in the query slot of a FlashReportQuery
object, and return the results in the raw_results slot.
}
\section{Methods (by class)}{
\itemize{
\item \code{FlashReportQuery}: Run the query slot against the redshift database and
return the results.
}}

