% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.r
\docType{data}
\name{subaggregates}
\alias{subaggregates}
\title{A string containing all (user_group, subaggregate) pairs.}
\format{A length-one character vector.}
\usage{
subaggregates
}
\description{
A string containing all (user_group, subaggregate) pairs.
}
\keyword{datasets}

