% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.r
\name{substitute_dates}
\alias{substitute_dates}
\title{A generic function to substitute the correct dates into the query_prototype
slot of a FlashReportQuery object, and return the result in the query slot.}
\usage{
substitute_dates(frq)
}
\arguments{
\item{frq}{A FlashReportQuery object.}
}
\value{
A FlashReportQuery object.
}
\description{
A generic function to substitute the correct dates into the query_prototype
slot of a FlashReportQuery object, and return the result in the query slot.
}

