% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.r
\docType{methods}
\name{substitute_user_group_name}
\alias{substitute_user_group_name}
\alias{substitute_user_group_name,FlashReportQuery-method}
\title{A generic function to substitute the correct user_group_query and
user_group_name into the query slot of a FlashReportQuery object
, and return the result in the query slot.}
\usage{
substitute_user_group_name(frq)

\S4method{substitute_user_group_name}{FlashReportQuery}(frq)
}
\arguments{
\item{frq}{A FlashReportQuery object.}
}
\value{
A FlashReportQuery object.
}
\description{
A generic function to substitute the correct user_group_query and
user_group_name into the query slot of a FlashReportQuery object
, and return the result in the query slot.
}
\section{Methods (by class)}{
\itemize{
\item \code{FlashReportQuery}: Substitute the user_group and 
user_group_names into the query slot. Return result in the query slot.
}}

