% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.r
\name{summarise_by_subaggregate}
\alias{summarise_by_subaggregate}
\title{Summarises results by subaggregate.}
\usage{
summarise_by_subaggregate(long_flash_report_2,
  subaggregate_df = flashreport::subaggregates)
}
\arguments{
\item{long_flash_report_2}{The result of calling curate_user_groups.}

\item{subaggregate_df}{A data frame that matches each user group to an
appropriate subaggregate (content champions, enterprise champions, other
champions, and internal users).}
}
\value{
A summarised version of long_flash_report_2. Same column structure,
but fewer rows.
}
\description{
Summarises results by subaggregate.
}

