% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.r
\name{summarise_in_aggregate}
\alias{summarise_in_aggregate}
\title{Summarises results in aggregate.}
\usage{
summarise_in_aggregate(long_flash_report_2)
}
\arguments{
\item{long_flash_report_2}{The result of calling curate_user_groups.}
}
\value{
A summarised version of long_flash_report_2. Same column structure,
but fewer rows.
}
\description{
Summarises results in aggregate.
}

