% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.r
\name{format_raw_results}
\alias{format_raw_results}
\title{A generic function to format the raw_results slot of a FlashReportQuery
object into a standardized format that can be rbinded with all other
results.}
\usage{
format_raw_results(frq)
}
\arguments{
\item{frq}{An object that inherits from FlashReportQuery.}
}
\value{
A FlashReportQuery object of the same subtype that was input. The
final_results slot contains a data.frame with columns (user_group, pa_group,
date_range, variable, value).
}
\description{
A generic function to format the raw_results slot of a FlashReportQuery
object into a standardized format that can be rbinded with all other
results.
}

