% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.r
\name{get_min_date}
\alias{get_min_date}
\title{A generic function to generate the correct min_date given the max_date and
range_type of an object that inherits from FlashReportQuery.}
\usage{
get_min_date(frq)
}
\arguments{
\item{frq}{An object that inherits from FlashReportQuery.}
}
\value{
A FlashReportQuery object of the same subtype that was input.
}
\description{
A generic function to generate the correct min_date given the max_date and
range_type of an object that inherits from FlashReportQuery.
}

