% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.r
\name{get_prototype}
\alias{get_prototype}
\title{A generic function to grab the correct query prototype and stick it into the
query_prototype slot of a FlashReport query.}
\usage{
get_prototype(frq)
}
\arguments{
\item{frq}{An object that inherits from FlashReportQuery.}
}
\value{
A FlashReportQuery object of the same subtype that was input.
}
\description{
A generic function to grab the correct query prototype and stick it into the
query_prototype slot of a FlashReport query.
}

