% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.r
\name{get_results}
\alias{get_results}
\title{A function to perform all queries and dump results into a long data.frame.}
\usage{
get_results(date_ranges, query_types)
}
\arguments{
\item{date_ranges}{A data frame with columns range_types, max_dates where
each row represents a date range.}

\item{query_types}{A character vector describing the types of queries to
run. Appropriate entries correspond to the names of S4 query type classes
defined in R/query_functions.r (currently 'auQuery', 'paQuery',
'notificationsQuery' are the only valid options).}
}
\value{
A data frame with columns user_group (character), date_range (date
ending), variable (character), value (numeric).
}
\description{
A function to perform all queries and dump results into a long data.frame.
}

