% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.r
\docType{class}
\name{notificationsQuery-class}
\alias{notificationsQuery-class}
\title{An S4 class to represent a 'notifications' Flash Report query.}
\description{
An S4 class to represent a 'notifications' Flash Report query.
}
\section{Slots}{

\describe{
\item{\code{range_type}}{A character indicating whether to use a year-to-date range
'ytd', or a weekly range, 'week'.}
}}

