% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.r
\docType{class}
\name{paQuery-class}
\alias{paQuery-class}
\title{An S4 class to represent a 'platform action' Flash Report query.}
\description{
An S4 class to represent a 'platform action' Flash Report query.
}
\section{Slots}{

\describe{
\item{\code{dummy}}{A dummy slot to differentiate this class from the
FlashReportQuery class.}
}}

