% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.r
\name{run_query}
\alias{run_query}
\title{A generic function to run the query in the query slot of a FlashReportQuery
object, and return the results in the raw_results slot.}
\usage{
run_query(frq, ...)
}
\arguments{
\item{frq}{A FlashReportQuery object.}

\item{...}{Placeholder for database connection object.}
}
\value{
A FlashReportQuery object.
}
\description{
A generic function to run the query in the query slot of a FlashReportQuery
object, and return the results in the raw_results slot.
}

