% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.r
\docType{methods}
\name{substitute_dates}
\alias{substitute_dates}
\alias{substitute_dates,FlashReportQuery-method}
\title{A generic function to substitute the correct dates into the query_prototype
slot of a FlashReportQuery object, and return the result in the query slot.}
\usage{
substitute_dates(frq)

\S4method{substitute_dates}{FlashReportQuery}(frq)
}
\arguments{
\item{frq}{A FlashReportQuery object.}
}
\value{
A FlashReportQuery object.
}
\description{
A generic function to substitute the correct dates into the query_prototype
slot of a FlashReportQuery object, and return the result in the query slot.
}
\section{Methods (by class)}{
\itemize{
\item \code{FlashReportQuery}: Substitute the min_date and max_date slots into
the query_prototype slot. Return result in the query slot.
}}

